feat: fix tests by removing trt dependency in tests
